+++ torch_geometric/data/hetero_data.py
@@ -502,7 +502,11 @@ class HeteroData(BaseData, FeatureStore, GraphStore):
         """
         return self.node_types, self.edge_types
 
-    def collect(self, key: str) -> Dict[NodeOrEdgeType, Any]:
+    def collect(
+        self,
+        key: str,
+        allow_empty: bool = False,
+    ) -> Dict[NodeOrEdgeType, Any]:
         r"""Collects the attribute :attr:`key` from all node and edge types.
 
         .. code-block:: python
@@ -517,12 +521,21 @@ class HeteroData(BaseData, FeatureStore, GraphStore):
         .. note::
 
             This is equivalent to writing :obj:`data.x_dict`.
+
+        Args:
+            key (str): The attribute to collect from all node and ege types.
+            allow_empty (bool, optional): If set to :obj:`True`, will not raise
+                an error in case the attribute does not exit in any node or
+                edge type. (default: :obj:`False`)
         """
         mapping = {}
         for subtype, store in chain(self._node_store_dict.items(),
                                     self._edge_store_dict.items()):
             if hasattr(store, key):
                 mapping[subtype] = getattr(store, key)
+        if not allow_empty and len(mapping) == 0:
+            raise KeyError(f"Tried to collect '{key}' but did not find any "
+                           f"occurrences of it in any node and/or edge type")
         return mapping
 
     def _check_type_name(self, name: str):
@@ -888,7 +901,7 @@ class HeteroData(BaseData, FeatureStore, GraphStore):
             value = torch.cat(values, dim) if len(values) > 1 else values[0]
             data[key] = value
 
-        if len(self.edge_label_index_dict) > 0:
+        if 'edge_label_index' in self:
             edge_label_index_dict = self.edge_label_index_dict
             for edge_type, edge_label_index in edge_label_index_dict.items():
                 edge_label_index = edge_label_index.clone()
@@ -1093,7 +1106,7 @@ def to_homogeneous_edge_index(
     cumsum = 0
     edge_indices: List[Tensor] = []
     edge_slices: Dict[EdgeType, Tuple[int, int]] = {}
-    for edge_type, edge_index in data.edge_index_dict.items():
+    for edge_type, edge_index in data.collect('edge_index', True).items():
         edge_index = offset_edge_index(node_slices, edge_type, edge_index)
         edge_indices.append(edge_index)
         edge_slices[edge_type] = (cumsum, cumsum + edge_index.size(1))